JlrsReflect.jl Documentation

One of the features of jlrs, a crate that provides bindings to the Julia C API for the Rust programming language, is that it supports mapping a struct from Julia to a Rust struct which lets you access its contents directly. Writing down the correct implementation of these structs in Rust can be challenging, this package can be used to automate the process.

Many types are supported, the major exceptions are types with tuple or union fields that depend on a free type parameter. These fields can have very different representations depending on these parameters which makes generating an appropriate mapping impossible. Everything else is fine, though; Types with bits union fields, UnionAll fields, type parameters, and "special" types backed by a struct defined in C like Module and Array are all supported.

There is a single entrypoint, the reflect method which takes a Vector of Types. Mappings will be generated for all of these types and their dependencies. The result of a call to this method can be written to a .rs-file and used as a module without any tweaking.

This package is compatible with Julia v1.5 and jlrs v0.6.

Main.JlrsReflect.reflect — Function

reflect(types::Vector{<:Type})::Bindings

Generate Rust mappings for all types in types and their dependencies. The only requirement is that these types must not contain any union or tuple fields that depend on a free type parameter. The mappings will derive JuliaStruct, and IntoJulia if it's a bits type.

The result of this method can be written to a file, its contents will be a valid Rust module.

When you use these mappings with jlrs, these types must be available with the same path. For example, if you generate bindings for Main.Bar.Baz, they must be available on that path and not some other path like Main.Foo.Bar.Baz.

Example

julia> using JlrsReflect

julia> reflect([StackTraces.StackFrame])
#[repr(C)]
#[jlrs(julia_type = "Base.StackTraces.StackFrame")]
#[derive(Copy, Clone, Debug, JuliaStruct)]
pub struct StackFrame<'frame, 'data> {
    pub func: ::jlrs::value::symbol::Symbol<'frame>,
    pub file: ::jlrs::value::symbol::Symbol<'frame>,
    pub line: i64,
    pub linfo: ::jlrs::value::Value<'frame, 'data>,
    pub from_c: bool,
    pub inlined: bool,
    pub pointer: u64,
}

#[repr(C)]
#[jlrs(julia_type = "Core.CodeInfo")]
#[derive(Copy, Clone, Debug, JuliaStruct)]
pub struct CodeInfo<'frame, 'data> {
    pub code: ::jlrs::value::array::Array<'frame, 'data>,
    pub codelocs: ::jlrs::value::Value<'frame, 'data>,
    pub ssavaluetypes: ::jlrs::value::Value<'frame, 'data>,
    pub ssaflags: ::jlrs::value::array::Array<'frame, 'data>,
    pub method_for_inference_limit_heuristics: ::jlrs::value::Value<'frame, 'data>,
    pub linetable: ::jlrs::value::Value<'frame, 'data>,
    pub slotnames: ::jlrs::value::array::Array<'frame, 'data>,
    pub slotflags: ::jlrs::value::array::Array<'frame, 'data>,
    pub slottypes: ::jlrs::value::Value<'frame, 'data>,
    pub rettype: ::jlrs::value::Value<'frame, 'data>,
    pub parent: ::jlrs::value::Value<'frame, 'data>,
    pub edges: ::jlrs::value::Value<'frame, 'data>,
    pub min_world: u64,
    pub max_world: u64,
    pub inferred: bool,
    pub inlineable: bool,
    pub propagate_inbounds: bool,
    pub pure: bool,
}

#[repr(C)]
#[jlrs(julia_type = "Core.Nothing")]
#[derive(Copy, Clone, Debug, JuliaStruct)]
pub struct Nothing {
}